+++ ESPixelStick/src/output/OutputSerial.cpp
@@ -136,7 +136,7 @@ void c_OutputSerial::GetDriverName(String &sDriverName)
 } // GetDriverName
 
 //----------------------------------------------------------------------------
-void c_OutputSerial::SetOutputBufferSize(size_t NumChannelsAvailable)
+void c_OutputSerial::SetOutputBufferSize(uint32_t NumChannelsAvailable)
 {
     // DEBUG_START;
     // DEBUG_V (String ("NumChannelsAvailable: ") + String (NumChannelsAvailable));
@@ -384,7 +384,7 @@ uint32_t IRAM_ATTR c_OutputSerial::ISR_GetNextIntensityToSend ()
         case SerialFrameState_t::RenardSendEscapedData:
         {
             data = *NextIntensityToSend - uint8_t(RenardFrameDefinitions_t::ESCAPED_OFFSET);
-            ++NextIntensityToSend;           
+            ++NextIntensityToSend;
             if (0 == --intensity_count)
             {
                 SerialFrameState = SerialFrameState_t::SerialIdle;
