+++ ESPixelStick/src/output/OutputUCS1903.cpp
@@ -79,7 +79,7 @@ void c_OutputUCS1903::GetStatus (ArduinoJson::JsonObject& jsonStatus)
 } // GetStatus
 
 //----------------------------------------------------------------------------
-void c_OutputUCS1903::SetOutputBufferSize (uint16_t NumChannelsAvailable)
+void c_OutputUCS1903::SetOutputBufferSize (uint32_t NumChannelsAvailable)
 {
     // DEBUG_START;
