+++ ESPixelStick/src/output/OutputUCS1903Rmt.cpp
@@ -120,7 +120,7 @@ bool c_OutputUCS1903Rmt::SetConfig (ArduinoJson::JsonObject& jsonConfig)
 } // SetConfig
 
 //----------------------------------------------------------------------------
-void c_OutputUCS1903Rmt::SetOutputBufferSize (uint16_t NumChannelsAvailable)
+void c_OutputUCS1903Rmt::SetOutputBufferSize (uint32_t NumChannelsAvailable)
 {
     // DEBUG_START;
